+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Added:
+- Crafting rock recipe data system
 - Dehydration water compatibility
 ### Fixed:
 -